+++ web/libs/datamanager/src/components/Common/Table/Table.scss
