+++ web/libs/datamanager/src/components/Common/Table/TableHead/TableHead.scss
