+++ web/libs/datamanager/src/components/Common/Table/TableRow/TableRow.scss
