+++ web/libs/datamanager/src/components/MainView/GridView/GridView.jsx
@@ -6,7 +6,7 @@ import InfiniteLoader from "react-window-infinite-loader";
 import { Block, Elem } from "../../../utils/bem";
 import { Checkbox } from "../../Common/Checkbox/Checkbox";
 import { Space } from "../../Common/Space/Space";
-import { getProperty, prepareColumns } from "../../Common/TableOld/utils";
+import { getProperty, prepareColumns } from "../../Common/Table/utils";
 import * as DataGroups from "../../DataGroups";
 import "./GridView.scss";
 import { FF_LOPS_E_3, isFF } from "../../../utils/feature-flags";
